fix(#83): fixed role deletion

+++ backend/api/interfaces/api_controllers/role_endpoint.py
@@ -1,11 +1,14 @@
 from typing import Annotated, List
+from beanie import PydanticObjectId
 from fastapi import APIRouter, Depends, status
+from api.common.exceptions import InvalidOperationException
 from api.common.utils import get_logger
 from api.domain.dtos.role_dto import CreateRoleDto, CreateRoleResponseDto, RoleDto, RoleListDto, UpdateRoleDto
-from api.core.container import get_role_service
+from api.core.container import get_role_service, get_user_service
 from api.domain.enum.permission import Permission
 from api.interfaces.security.role_checker import check_permissions_for_current_role
 from api.usecases.role_service import RoleService
+from api.usecases.user_service import UserService
 
 
 logger = get_logger(__name__)
@@ -65,12 +68,19 @@ async def update_role(
     return RoleDto(**serialized_role)
 
 
-@router.delete("/{role_id}", status_code=status.HTTP_202_ACCEPTED)
+@router.delete("/{role_id:path}", status_code=status.HTTP_202_ACCEPTED, description="Deletes a role if it is not assigned to any users.")
 async def delete_role(
     role_id: str,
     service: RoleService = Depends(get_role_service),
+    user_service: UserService = Depends(get_user_service),
     _bool: bool = Depends(check_permissions_for_current_role(required_permissions=[Permission.ROLE_DELETE_ONLY]))
 ):
+    users_role = await user_service.total_count(params={"role_id": role_id})
+    logger.debug(f"Number of users with role {role_id}: {users_role}")
+    if users_role > 0:
+        logger.warning(f"Cannot delete role {role_id} as it is assigned to users.")
+        raise InvalidOperationException(f"Cannot delete role {role_id} as it is assigned to users.")
+    
     await service.delete_role(role_id)
     return status.HTTP_202_ACCEPTED
